+++ CHANGES.rst
@@ -34,6 +34,8 @@ astropy.io.misc
 
 - Add implementations of astropy-specific ASDF tag types. [#6790]
 
+- Add ASDF tag and schema for ICRSCoord. [#6904]
+
 astropy.io.fits
 ^^^^^^^^^^^^^^^